+++ src/main/java/br/com/carloslonghi/eletrolonghi/controller/DeviceController.java
@@ -11,6 +11,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/device")
@@ -60,9 +61,14 @@ public class DeviceController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteDeviceById(@PathVariable Long id) {
-        deviceService.deleteById(id);
+        Optional<Device> optionalDevice = deviceService.findById(id);
 
-        return ResponseEntity.noContent().build();
+        if (optionalDevice.isPresent()) {
+            deviceService.deleteById(id);
+            return ResponseEntity.noContent().build();
+        }
+
+        return ResponseEntity.notFound().build();
     }
 
     @GetMapping("/search")
